fix: bind the timeline wheel handler regardless of audio and lyrics load order
